api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
redis==5.0.8
